[SDAN-661] Use Newsroom.Resource|Service for resource classes (#997)
* [SDAN-661] Use Newsroom.Resource|Service for resource classes
* Fixed monitoring server unit tests
