+++ newsroom/news_api/formatters/service.py
@@ -1,6 +1,5 @@
 from newsroom.wire.formatters import get_all_formatters
 from superdesk.utils import ListCursor
-from superdesk.services import BaseService
 from flask import abort
 from superdesk import get_resource_service
 from flask import current_app as app
@@ -8,9 +7,10 @@ from eve.versioning import versioned_id_field
 from datetime import timedelta
 from superdesk.utc import utcnow
 from newsroom.settings import get_setting
+from newsroom import Service
 
 
-class APIFormattersService(BaseService):
+class APIFormattersService(Service):
     """
     Internal service the will format the requested item in the requested format
     """
